Started implementing user registration.
Alas, we have reached a turning point in our development process. As we
finally begin to host our beloved TunnelMessenger online, we need to
start dealing with users. So, I have started working on the user account
registration feature of our software.

At the point of this commit, I merely started creating a very basic
registration page, as well as parsing the data from the formulary
and verifying if passwords match and if the e-mail is properly
formatted.