Implemented login validation and redirect
Oh my, what a coincidence to see you here. How do you fare
today? I have implemented basic login funcionality into
the server, but it isn't an actual login yet. All it
does is verify if the input username and password are
indeed in the database.

Passwords are still stored in plain text for now, which
we need to fix. Also, you can completely bypass the
login by adding /TunnelMessenger on the address, so
for now, the login is useless.