Chess is coming soon
Howdy, dear developers of TunnelMessenger. It is my honor to announce
that an exciting feature is coming soon to TunnelMessenger. That would
be nothing more, nothing less than FUCKING CHESS.

In order to reflect such an amazing feature, I have included a humble,
but sincere banner in our login page.