Started implementing Magic The Gathering commands
Hello everyone. One thing that I always do when I am talking
with fellow MTG fans, is to find card links and images from
the Scryfall or Gatherer databases and send them on telegram
or whatsapp. I always thought it would be great if my messenger
had a builtin feature to search and share MTG cards. SO
I DECIDED TO ADD IT TO TUNNEL MESSENGER.

In this commit, I begin adding the necessary structures to
make the /mtg command a reality. Happy Tunneling!